+++ docs/pages/release_notes.rst
@@ -99,7 +99,7 @@ The HTTP interface has been reworked to make deployments simpler and more secure
 - No more dynamic binding of endpoints to ports, endpoints are now bound to the internal interface (``8081``) or the public interface (``8080``).
 - Server-side TLS for HTTP has been dropped, since the Nuts node is always expected to be deployed behind a reverse proxy/ingress that handles TLS termination.
 - API authentication is now only applied to ``/internal`` endpoints, since those are the only API endpoints that should be protected with authentication.
-- CORS support has been removed. As it is only required by user authentication endpoints that are considered to be deprecated, CORS headers can be set by a reverse proxy if still required.
+- CORS configuration for IRMA/Yivi has been moved to the `auth.irma.cors.origin` config parameter.
 
 Port configuration
 ------------------